requirements.txt

# Trigger rebuildstreamlit
jinja2

python-docx==1.2.0
scikit-learn
weasyprint==67.0
